Changed a lot
i dont remember all the stuff i changed. Mainly made it so main tracker is running without any blocking calls to tracked processes. Currently, if a tracked car parks, its process goes "not responding" because we stop tracking it the normal way and stop sending it ok messages to read the next frame. It should be updated to track when it leaves on another thread.