Updates
Made the new entrant detection run on a time delay instead of as many times as it could per second.

Set up the object broker with 2 legs. Need to figure out a new way of calculating where the new entrant came from as the current way is not reliable. Thinking of waiting for the next frame to figure out direction of movement and use that.